refactor: Group 패키지에 Participant 패키지 통합
