+++ backend/src/main/java/com/woowacourse/momo/group/service/GroupFindService.java
@@ -13,8 +13,8 @@ import lombok.RequiredArgsConstructor;
 
 import com.woowacourse.momo.global.exception.exception.ErrorCode;
 import com.woowacourse.momo.global.exception.exception.MomoException;
-import com.woowacourse.momo.group.domain.group.Group;
-import com.woowacourse.momo.group.domain.group.GroupRepository;
+import com.woowacourse.momo.group.domain.Group;
+import com.woowacourse.momo.group.domain.GroupRepository;
 import com.woowacourse.momo.group.service.dto.request.GroupFindRequest;
 import com.woowacourse.momo.group.service.specification.GroupSpecification;
 import com.woowacourse.momo.member.domain.Member;
