+++ backend/src/main/java/com/woowacourse/momo/group/service/specification/GroupAttribute.java
@@ -6,7 +6,7 @@ import javax.persistence.criteria.Expression;
 import javax.persistence.criteria.Path;
 import javax.persistence.criteria.Root;
 
-import com.woowacourse.momo.group.domain.group.Group;
+import com.woowacourse.momo.group.domain.Group;
 
 public enum GroupAttribute {
