+++ backend/src/main/java/com/woowacourse/momo/group/service/specification/GroupSpecification.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Component;
 
 import com.woowacourse.momo.category.domain.Category;
-import com.woowacourse.momo.group.domain.group.Group;
+import com.woowacourse.momo.group.domain.Group;
 import com.woowacourse.momo.member.domain.Member;
 
 @Component
