+++ backend/src/test/java/com/woowacourse/momo/group/controller/ParticipantControllerTest.java
@@ -1,4 +1,4 @@
-package com.woowacourse.momo.participant.controller;
+package com.woowacourse.momo.group.controller;
 
 import static org.hamcrest.Matchers.is;
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
@@ -35,17 +35,17 @@ import org.springframework.transaction.annotation.Transactional;
 import com.woowacourse.momo.auth.service.AuthService;
 import com.woowacourse.momo.auth.service.dto.request.LoginRequest;
 import com.woowacourse.momo.auth.service.dto.request.SignUpRequest;
+import com.woowacourse.momo.group.domain.Group;
 import com.woowacourse.momo.group.domain.calendar.Calendar;
 import com.woowacourse.momo.group.domain.calendar.Deadline;
 import com.woowacourse.momo.group.domain.calendar.Schedules;
-import com.woowacourse.momo.group.domain.group.Group;
 import com.woowacourse.momo.group.service.GroupFindService;
 import com.woowacourse.momo.group.service.GroupService;
+import com.woowacourse.momo.group.service.ParticipantService;
 import com.woowacourse.momo.group.service.dto.request.DurationRequest;
 import com.woowacourse.momo.group.service.dto.request.GroupRequest;
 import com.woowacourse.momo.group.service.dto.request.ScheduleRequest;
 import com.woowacourse.momo.member.service.MemberService;
-import com.woowacourse.momo.participant.service.ParticipantService;
 
 @AutoConfigureMockMvc
 @AutoConfigureRestDocs
