+++ backend/src/test/java/com/woowacourse/momo/group/domain/participant/CapacityRangeTest.java
@@ -1,4 +1,4 @@
-package com.woowacourse.momo.group.domain.group;
+package com.woowacourse.momo.group.domain.participant;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
@@ -8,13 +8,13 @@ import org.junit.jupiter.params.provider.ValueSource;
 
 import com.woowacourse.momo.global.exception.exception.MomoException;
 
-class GroupCapacityRangeTest {
+class CapacityRangeTest {
 
     @DisplayName("정원의 수가 범위를 벗어나면 예외가 발생한다")
     @ParameterizedTest
     @ValueSource(ints = {-1, 0, 100})
     void validateCapacityIsInRange(int capacity) {
-        assertThatThrownBy(() -> GroupCapacityRange.validateCapacityIsInRange(capacity))
+        assertThatThrownBy(() -> CapacityRange.validateCapacityIsInRange(capacity))
                 .isInstanceOf(MomoException.class)
                 .hasMessage("모임 내 인원은 1명 이상 99명 이하여야 합니다.");
     }
